backend/requirements.txt
# Core Django and REST API
Django<5.0,>=4.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0

# Caching
django-redis==5.4.0
redis==5.0.1
hiredis==2.3.2

# Monitoring
django-prometheus==2.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-yasg==1.21.7

# HTTP Requests
requests==2.31.0

# Development Tools
black==24.3.0
pathspec==0.11.2
